// contracts/generated/typescript/models/ClaimDetail.ts
export interface ClaimDetail {
    /**
     * 
     * @type {string}
     * @memberof ClaimDetail
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ClaimDetail
     */
    subjectEntityId: string;
    /**
     * 
     * @type {PropertySummary}
     * @memberof ClaimDetail
     */
    property: PropertySummary;
    /**
     * 
     * @type {string}
     * @memberof ClaimDetail
     */
    valueType: string;
    /**
     * 
     * @type {any}
     * @memberof ClaimDetail
     */
    value: any | null;
    /**
     * 
     * @type {string}
     * @memberof ClaimDetail
     */
    targetEntityId: string;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof ClaimDetail
     */
    qualifiers: { [key: string]: any; };
    /**
     * 
     * @type {ClaimDetailRankEnum}
     * @memberof ClaimDetail
     */
    rank: ClaimDetailRankEnum;
    /**
     * 
     * @type {ClaimDetailStatusEnum}
     * @memberof ClaimDetail
     */
    status: ClaimDetailStatusEnum;
    /**
     * 
     * @type {ClaimDetailVerificationStatusEnum}
     * @memberof ClaimDetail
     */
    verificationStatus: ClaimDetailVerificationStatusEnum;
    /**
     * 
     * @type {Date}
     * @memberof ClaimDetail
     */
    validFrom: Date;
    /**
     * 
     * @type {Date}
     * @memberof ClaimDetail
     */
    validTo: Date;
    /**
     * 
     * @type {ClaimDetailOriginTypeEnum}
     * @memberof ClaimDetail
     */
    originType: ClaimDetailOriginTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof ClaimDetail
     */
    createdBy: string;
    /**
     * 
     * @type {Date}
     * @memberof ClaimDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof ClaimDetail
     */
    supersededBy: string;
    /**
     * 
     * @type {Array<ClaimSource>}
     * @memberof ClaimDetail
     */
    sources: Array<ClaimSource>;
}


/**
 * @export
 */
export const ClaimDetailRankEnum = {
    Preferred: 'preferred',
    Normal: 'normal',
    Deprecated: 'deprecated'
} as const;
export type ClaimDetailRankEnum = typeof ClaimDetailRankEnum[keyof typeof ClaimDetailRankEnum];

/**
 * @export
 */
export const ClaimDetailStatusEnum = {
    Proposed: 'proposed',
    Published: 'published',
    Rejected: 'rejected',
    Superseded: 'superseded',
    Deprecated: 'deprecated'
} as const;
export type ClaimDetailStatusEnum = typeof ClaimDetailStatusEnum[keyof typeof ClaimDetailStatusEnum];

/**
 * @export
 */
export const ClaimDetailVerificationStatusEnum = {
    Unverified: 'unverified',
    AiChecked: 'ai_checked',
    HumanVerified: 'human_verified',
    Disputed: 'disputed'
} as const;
export type ClaimDetailVerificationStatusEnum = typeof ClaimDetailVerificationStatusEnum[keyof typeof ClaimDetailVerificationStatusEnum];

/**
 * @export
 */
export const ClaimDetailOriginTypeEnum = {
    Human: 'human',
    Ai: 'ai',
    Import: 'import'
} as const;
export type ClaimDetailOriginTypeEnum = typeof ClaimDetailOriginTypeEnum[keyof typeof ClaimDetailOriginTypeEnum];